Refactor validators
Create GenomicEntity validator that is extended by AGM validator.
Also added basic validation (regex) for taxon IDs - this can be
replaced with a DB lookup once taxons are loaded.
